Initial configuration for neovim
This contains the initial settings for my neovim configuration. While
this may be an incomplete feature set, it contains the foundational
pieces.

I have a goal of setting up isolated environments depending on the
development that I am aiming to do.

For example, you will notice a "global" configuration directory. This
directory contains all the universally shareable settings.

There is also a "standard" configuration directory. This directory
contains the standard plugins and extensions that I find useful in many
scenarios. Things like oil and telescope. This configuration will be the
"default" neovim configuration when I just run 'nvim' without specifying
an environment.

From here, the intention is to build out a "ruby on rails" environment
and a "vue" environment and a python environment, etc. Each will be
configured slightly differently, but the point is to keep the launch of
neovim as quick and unbloated as possible depending on what I am aiming
to accomplish during that development session.